refactor: split report integration test into direct-only and multi-EDP variants
Split InProcessLifeOfAReportIntegrationTest (~3277 lines) into a base
class with shared infrastructure and three focused test targets:
DirectOnly (27 single-EDP tests), HMSS MultiEdp (2 cross-publisher
tests with 3 duchies), and TrusTee MultiEdp (2 cross-publisher tests
with 1 duchy). Add trusTeeEnabled parameter to InProcessCmmsComponents
and InProcessEdpSimulator for protocol selection control.

Issue: #3820
