+++ src/test/kotlin/org/wfanet/measurement/integration/deploy/gcloud/GCloudInProcessTrusTeeMultiEdpReportIntegrationTest.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 The Cross-Media Measurement Authors
+ * Copyright 2026 The Cross-Media Measurement Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,23 +20,26 @@ import org.junit.ClassRule
 import org.wfanet.measurement.common.db.r2dbc.postgres.testing.PostgresDatabaseProviderRule
 import org.wfanet.measurement.duchy.deploy.common.postgres.testing.Schemata.DUCHY_CHANGELOG_PATH
 import org.wfanet.measurement.gcloud.spanner.testing.SpannerEmulatorRule
-import org.wfanet.measurement.integration.common.ALL_DUCHY_NAMES
+import org.wfanet.measurement.integration.common.AGGREGATOR_NAME
 import org.wfanet.measurement.integration.common.IMPRESSION_QUALIFICATION_FILTER_MAPPING
-import org.wfanet.measurement.integration.common.reporting.v2.InProcessLifeOfAReportIntegrationTest
+import org.wfanet.measurement.integration.common.reporting.v2.InProcessMultiEdpReportIntegrationTest
 import org.wfanet.measurement.integration.deploy.common.postgres.PostgresDuchyDependencyProviderRule
 import org.wfanet.measurement.reporting.deploy.v2.postgres.testing.Schemata.REPORTING_CHANGELOG_PATH as POSTGRES_REPORTING_CHANGELOG_PATH
 
-/** Implementation of [InProcessLifeOfAReportIntegrationTest] for Google Cloud. */
-class GCloudInProcessLifeOfAReportV2IntegrationTest :
-  InProcessLifeOfAReportIntegrationTest(
+/** TrusTee implementation of [InProcessMultiEdpReportIntegrationTest] with 1 duchy. */
+class GCloudInProcessTrusTeeMultiEdpReportIntegrationTest :
+  InProcessMultiEdpReportIntegrationTest(
     KingdomDataServicesProviderRule(spannerEmulator),
-    PostgresDuchyDependencyProviderRule(duchyDatabaseProvider, ALL_DUCHY_NAMES),
+    PostgresDuchyDependencyProviderRule(duchyDatabaseProvider, listOf(AGGREGATOR_NAME)),
     SpannerAccessServicesFactory(spannerEmulator),
     InternalReportingServicesProviderRule(
       spannerEmulator,
       reportingPostgresDatabaseProvider,
       IMPRESSION_QUALIFICATION_FILTER_MAPPING,
     ),
+    duchyNames = listOf(AGGREGATOR_NAME),
+    hmssEnabled = false,
+    trusTeeEnabled = true,
   ) {
   companion object {
     @get:ClassRule @JvmStatic val spannerEmulator = SpannerEmulatorRule()
